requirements.txt
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
fastapi==0.97.0
fastapi-utils==0.2.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
pycparser==2.21
pydantic==1.10.9
PyJWT==2.7.0
pytest==7.3.2
pytest-asyncio==0.21.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==1.4.48
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
